+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ v0.2.4 (????-??-??)
 * added `persist` method to `Container`
 * added `pid` property `Container`
 * added `_info` property to `Container`
+* added `find` method to `FileSystem`
 
 
 v0.2.3 (2020-01-22)